src/requirements.txt
anyascii==0.3.2
asgiref==3.8.1
beautifulsoup4==4.11.2
black==24.4.2
boto3==1.34.109
botocore==1.34.109
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coverage==7.5.1
defusedxml==0.7.1
Django==5.0.1
django-debug-toolbar==4.3.0
django-filter==23.5
django-modelcluster==6.3
django-permissionedforms==0.1
django-storages==1.14.3
django-stubs==4.2.7
django-stubs-ext==5.0.0
django-taggit==4.0.0
django-treebeard==4.7.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==3.14.5
django-cors-headers
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
factory-boy==3.3.0
Faker==25.2.0
filetype==1.2.0
gevent==23.9.1
greenlet==3.0.3
gunicorn==21.2.0
html5lib==1.1
idna==3.7
iniconfig==2.0.0
isort==5.13.2
jmespath==1.0.1
l18n==2021.3
mypy==1.8.0
mypy-extensions==1.0.0
openpyxl==3.1.2
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
pillow_heif==0.16.0
pipdeptree==2.21.0
platformdirs==4.2.2
pluggy==1.5.0
psycopg==3.1.17
ptvsd==4.3.2
PyJWT==2.8.0
pytest==8.0.0
pytest-django==4.8.0
pytest-env==1.1.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.1
ruff==0.4.4
s3transfer==0.10.1
sentry-sdk==1.39.2
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
telepath==0.3.1
types-pytz==2024.1.0.20240417
types-PyYAML==6.0.12.20240311
types-requests==2.32.0.20240521
typing_extensions==4.11.0
urllib3==2.2.1
wagtail==5.2.3
wagtail-factories==4.1.0
wagtail-headless-preview==0.7.0
wagtail-meta-preview==4.0.0
webencodings==0.5.1
whitenoise==6.6.0
Willow==1.6.3
zope.event==5.0
zope.interface==6.4
drf_social_oauth2==2.2.0
